fixing capitalization when including files
mingw ships with dshow.h, aviriff.h and windows.h.

however, if you use it to cross-compile on a linux-host, chances are
high, that the filesystem is actually case sensitive (e.g. ext3).

if this is the case "#include <DShow.h>" will fail
(as the file "DShow.h" is not the same as "dshow.h")