const'ify argument to getDeviceIDFromName
when using C++, we often deal with std::string rather than "char*"

thus, the device-name will often be passed as "s.c_str()", which really
is a "const char*".

to avoid warnings, it would be good if videoInput::getDeviceIDFromName()
would accept "const char*".

note: this probably breaks ABI compatibility!
in this case it might be better to overload this member:
  int videoInput::getDeviceIDFromName(const std::string&name);